Create Basic Rules
Make sure your data and R Markdown files are in the same directory. If not, go to Session in the RStudio menu bar, select 'Set Working Directory' and 'Choose Directory' as the same folder where you have your data. You can also use the 'setwd()' command in the console and specify the path.
ex: 
>setwd("/folder/sub-folder/sub-sub-folder")